mqtt/producer/image_requirements.txt
# Image processing dependencies for AirTracker
Pillow>=10.0.0
requests>=2.25.0
python-dotenv>=1.0.0